Find vcpkg bootstrap in the current directory
If `AutoVcpkg.cmake` and `vcpkg-bootstrap.cmake` are not in the same
directory as `CMakeLists.txt`, the script cannot find
`vcpkg-bootstrap.cmake`. Assuming that both `AutoVcpkg.cmake` and
`vcpkg-bootstrap.cmake` are always in the same directory, they both can
be found in the directory indicated by `CMAKE_CURRENT_FUNCTION_LIST_DIR`
variable. This requires at least CMake 3.17.